feat(docs): autogenerating cli references spike
This defines both the template, and the method.

The template is simple, and uses one page.

At the entrypoint we have `aztec`

#command (aztec)
## table of contents
## command detail 1
### usage
### description
### argument
### options
## command detail 2...

note that aztec start is a wonky example

it lives on a command detail level, but maybe we should consider refactoring it to something like

aztec node
aztec sequencer

instead of the current `aztec start --node` `aztec start --sequencer`

This so we can extend the current model across aztec start

This PR adds scripts to generate this and gives an example of the an autogenerated file.

Co-authored-by: Esau <esau@aztecprotocol.com>