feat: introducing #[internal]
Fixes F-22
Fixes #9180

Introduces `#[internal(...)]` macro that injects ContractSelf just like `#[external(...)]` macro but instead of constructing relevant context from inputs it directly accepts the context as first arg to the function (this is also injected into params by the macro). Then when a call to internal is formed via `self.internal.my_internal_func(...)` the call directly passes the relevant context.

The same approach is used as what was introduced for`#[external(...)]` in a PR down the stack in that the `#[aztec]` macro makes the original functions uncallable and generates new function with `__aztec_nr_internals__` prefix and injects it into the contract. This gives devs a nice error if they attempt to call the function directly instead of via the intended API.