refactor: cleaning up flow of calls generation
Before this PR we generate the external function calls when the external macro was run and then we added the result to registries for later to be obtained by the aztec macro.

This flow became unnecessarily complex because now the external macro just registers the function definitions in the external function registry and then it uses this registry to generate the `__aztec_nr_internals__` functions in the aztec macro. We can use this same function registry also for generating the calls obviating the need for the call registries.