feat: add C api for freeing bb results
There does not appear to be any mechanism for freeing the results
returned by bbapi. Even though the memory is allocated by the standard C
allocator, it's safer to provide a destroy API that guarantees that the
same allocator that allocated the memory is also used to free the memory.