refactor: restructuring function macros + migration notes
In the past core of the macro logic was handled by the external macro code. This has changed and now it's all done by Aztec macro.

This resulted in a lot of the macro functionality now being placed in a way that made no sense. In the PR diff I comment on where each function was moved to.

I also update migration notes and I describe there all the new contract API.