feat: noir compilation and witness generation api (#8)
* feat: nargo api- compile workspace

Adds a function that compiles a workspace into a ProgramArtifact
instance

* feat: execute contract

Adds an API function to run a noir ProgramArtifact with a set of inputs
to create a witness.

* chore: remove remaining black-box functions

Removes the remaining native crypto functions from the library. bb_rs is
meant to be a circuit proving and verifying API.

* chore: clean up yagni code

Removes unused code.

* fix: call bbapi_clenup

Main features
---

compile: Adds a function that compiles a workspace into a ProgramArtifact instance

execute: Adds an API function to run a noir ProgramArtifact with a set of inputs to create a witness.


Cleanup
---
- Removes non-ZK functions (e.g. Schnorr signatures, ECDSA, hashing etc) from the bb_rs API. The only real interest for this library are the Snark provers and related utilities.
- It's important to call bbapi_cleanup for hygiene reasons. Right now it's ano-op, but might do actual cleanup in future, and we don't want to create a memory leak.