feat(zkuh-rs)!: use precompiled library (#9)
* feat(zkuh-rs)!: use precompiled library

I can't quite get the entire toolchain to compile automatically when
using zkuh-rs as a dependency, so I'm throwing in the towel and
committing the compiled asset to VCS.

You can compare the checksums.txt file to ensure that the libraries that
you have match what was built.

Or you can still run `bootstrap.sh build_barretenberg` to rebuild the
library and compare hashes.